+++ packages/khao-ui/src/components/forms/textfield/TextField.svelte
@@ -7,6 +7,7 @@
   } from "../types/TextFieldType";
   import Icon from "../../../icons/Icon.svelte";
   import { type IconName } from "../../../icons/types/IconName";
+  import type { StringBoolean } from "../../../common/types/StringBoolean";
 
   export let label: string;
   export let value: string;
@@ -14,9 +15,27 @@
   export let id: string = `khao-ui-texfield-${label}`;
   export let type: TextFieldType = textFieldTypeDefault;
   export let iconName: IconName | "" = "";
+  export let autofocus: StringBoolean = "false";
+  export let disabled: StringBoolean = "false";
 
   if (type === "search" && iconName === "") {
     iconName = "search";
+  }
+
+  function handleChange(event: Event) {
+    const input = event.target as HTMLInputElement;
+
+    const CHANGE_EVENT_NAME = "khao-text-field-change";
+
+    const changeEvent = new CustomEvent(CHANGE_EVENT_NAME, {
+      detail: {
+        value: input.value,
+      },
+      bubbles: true,
+      composed: true,
+    });
+
+    input.dispatchEvent(changeEvent);
   }
 </script>
 
@@ -30,7 +49,17 @@
   {#if type === "output"}
     <output class="field" {id}>{value}</output>
   {:else}
-    <input class="field" {placeholder} {id} {type} {value} />
+    <!-- svelte-ignore a11y-autofocus -->
+    <input
+      class="field"
+      {placeholder}
+      {id}
+      {type}
+      {value}
+      autofocus={autofocus === "true" ? true : false}
+      disabled={disabled === "true" ? true : false}
+      on:keydown={handleChange}
+    />
   {/if}
 </div>